backend/pyproject.toml
[tool.poetry]
name = "analinet"
version = "0.1.0"
description = "Sistema integral para la gestión de laboratorios clínicos con capacidades de IA"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
aiofiles = "^23.1.0"
reportlab = "^4.0.4"
python-barcode = "^0.14.0"
pyhl7 = "^1.5.0"
astm = "^0.7.0"
fhir-resources = "^6.5.0"
# Dependencias de IA
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
tensorflow = "^2.13.0"
joblib = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
